chapter_003/src/main/java/ru/job4j/list/SortUser.java
package ru.job4j.list;

import ru.job4j.user.User;

import java.util.Comparator;
import java.util.List;
import java.util.Set;
import java.util.TreeSet;

public class SortUser {

    public Set<User> sort(List<User> users) {
        return new TreeSet<>(users);
    }

    public List<User> sortNameLength(List<User> users) {
        users.sort(new Comparator<User>() {
            @Override
            public int compare(User o1, User o2) {
                final boolean rs1 = o1.getName().length() == o2.getName().length();
                return rs1 ? 0 : o1.getName().length() > o2.getName().length() ? 1 : -1;
            }
        });
        return users;
    }

    public List<User> sortByAllFields(List<User> users) {
        users.sort(new Comparator<User>() {
            @Override
            public int compare(User o1, User o2) {
                return o1.getName().compareTo(o2.getName());
            }
        });
        return users;
    }
}